Revert "refactor invariant campaign state"
This reverts commit 450f80f73fd2448496551c297ee4dc59be721c90.
